+++ src/Firebase/Auth/Token/Exception/UnknownKey.php
@@ -2,21 +2,23 @@
 
 namespace Firebase\Auth\Token\Exception;
 
-class UnknownKey extends \LogicException
+use Lcobucci\JWT\Token;
+
+class UnknownKey extends InvalidToken
 {
     /**
      * @var string
      */
     private $keyId;
 
-    public function __construct($keyId)
+    public function __construct(Token $token, string $keyId)
     {
-        parent::__construct(sprintf('A key with ID "%s" could not be found.', $keyId));
+        parent::__construct($token, sprintf('A key with ID "%s" could not be found.', $keyId));
 
         $this->keyId = $keyId;
     }
 
-    public function getKeyId()
+    public function getKeyId(): string
     {
         return $this->keyId;
     }