+++ src/JWT/Action/FetchGooglePublicKeys/Handler.php
@@ -6,8 +6,12 @@ namespace Kreait\Firebase\JWT\Action\FetchGooglePublicKeys;
 
 use Kreait\Firebase\JWT\Action\FetchGooglePublicKeys;
 use Kreait\Firebase\JWT\Contract\Keys;
+use Kreait\Firebase\JWT\Error\FetchingGooglePublicKeysFailed;
 
 interface Handler
 {
+    /**
+     * @throws FetchingGooglePublicKeysFailed
+     */
     public function handle(FetchGooglePublicKeys $action): Keys;
 }