+++ src/JWT/Contract/ExpirableTrait.php
@@ -31,12 +31,6 @@ trait ExpirableTrait
 
     public function expiresAt(): DateTimeImmutable
     {
-        // @codeCoverageIgnoreStart
-        if (!$this->expirationTime) {
-            throw new LogicException(static::class.' allows calling '.__METHOD__.' before setting it.');
-        }
-        // @codeCoverageIgnoreEnd
-
         return $this->expirationTime;
     }
 }